Running usm(s) from a/several directory(ies) or a/several subdirectory(ies) named with usm name

This function uses Stics directly through a system call

run_stics(
  stics_exe,
  workspace,
  usm = NULL,
  check = TRUE,
  verbose = FALSE,
  model_path = lifecycle::deprecated(),
  data_dir = lifecycle::deprecated(),
  usm_dir_names = lifecycle::deprecated(),
  check_exe = lifecycle::deprecated(),
  display = lifecycle::deprecated()
)

Arguments

stics_exe

Path of Stics the executable file

workspace

Path of the workspace containing the Stics (txt) input files

usm

Vector of USM names. Optional, if provided, the function runs only the given USMs. If not provided, the function runs all the USMs included in workspace.

check

Logical, T for checking the model executable, F otherwise

verbose

Logical value (optional), TRUE to display usms names, FALSE otherwise (default)

model_path

model_path is no longer supported, use stics_exe instead.

data_dir

data_dir is no longer supported, use workspace instead.

usm_dir_names

usm_dir_names is no longer supported, use usm instead.

check_exe

check_exe is no longer supported, use check instead.

display

display is no longer supported, use verbose instead.

Value

A list with usm names and execution error status

Examples

if (FALSE) {

# Specifying individual usm directories
run_stics("/home/username/bin/Stics", "/home/username/Work/SticsInputsDir")
run_stics("/home/username/bin/Stics", c(
  "/home/username/Work/SticsInputsDir1",
  "/home/username/Work/SticsInputsDir2"
))

# Specifying a parent directory of usms directories
# running one or several usms
run_stics("/home/username/bin/Stics",
"/home/username/Work/SticsInputsRootDir", "wheat")
run_stics(
  "/home/username/bin/Stics", "/home/username/Work/SticsInputsRootDir",
  c("wheat", "maize")
)
# running all usms
run_stics(
  "/home/username/bin/Stics", "/home/username/Work/SticsInputsRootDir",
)
}