Adding a new Stics model executable in JavaStics

Add a new executable of the Stics model to use in JavaStics, and set it as the one to use by default. Please refer to select_stics_exe() to change the stics executable used by JavaStics, and list_stics_exe() to list all available executables.

set_stics_exe(javastics_path, stics_exe, overwrite = FALSE, verbose = TRUE)

Arguments

javastics_path
JavaStics installation root folder

stics_exe
Stics executable name (identifier) or executable path

overwrite
Boolean. If stics_exe is an executable path and an executable with the same name already exist in the bin, overwrite it if TRUE, or return an error if FALSE default.

verbose
Logical value (optional), TRUE to display messages, FALSE otherwise.

Value

Nothing. Update the "preference.xml" file in the config of JavaStics.

Details

stics_exe may be :

a model name pointing to a stics executable as done in JavaStics, e.g. "modulostics" for stics_module.exe;

a stics executable file available from the bin folder in JavaStics, e.g. "stics_module.exe";

a path to a stics executable file, eg. "C:/Users/username/Desktop/stics.exe" The function will test those hypothesis in the same order they are presented in the paragraph above. Note that the stics executable will be copied in the bin folder of JavaStics in the third case, and named after the executable name and the user's OS, e.g. "stics_intercrop_win" for ane executable called "stics_intercrop.exe" in a windows OS. If the file already exists in the bin, the function will return an error overwrite= FALSE or will replace the executable if overwrite= TRUE. If the name already exist, it will check if the executable is the same. If it is, the same name is used, if not, a new name with an incremented index is given, e.g. stics_intercrop_win_2. In any case, the function will inform the user of which stics is being used to avoid any issue.

Note

"stics_modulo", "sticsmodulo" and "modulostics" are synonyms for the standard STICS executable.

Examples

if (FALSE) {
 # Using model name:
 set_stics_exe("/path/to/JavaSTICS/dir","modulostics")
 # Using model executable:
 set_stics_exe("/path/to/JavaSTICS/dir","stics_modulo.exe")
 # Using path to add a new executable:
 set_stics_exe("/path/to/JavaSTICS/dir","path/to/stics.exe")
}