Running usm(s) from txt input files stored in one directory per situation, simulated results are returned in a list

This function uses Stics directly through a system call, can force Stics input parameters with values given in arguments.

stics_wrapper(
  param_values = NULL,
  sit_var_dates_mask = NULL,
  prior_information = NULL,
  model_options
)

Arguments

param_values
named vector containing the value(s) and names of the parameters to force (optional)

sit_var_dates_mask
List of situations, variables and dates for which simulated values should be returned. Typically a list containing the observations to which simulations should be compared as provided by SticsRFiles::read_obs_to_list

prior_information
Prior information on the parameters to estimate. For the moment only uniform distribution are allowed. Either a list containing (named) vectors of upper and lower bounds (ub and lb), or a named list containing for each parameter the list of situations per group (sit_list) and the vector of upper and lower bounds (one value per group) (ub and lb)

model_options
List containing any information needed by the model. In the case of Stics: stics_path the path of Stics executable file and data_dir the path of the directory containing the Stics input data for each USM (one folder per USM where Stics input files are stored in txt format)

Value

A list containing simulated values (sim_list) and a flag (flag_allsim) indicating if all required situations, variables and dates were simulated.