Getting/setting a stics_wrapper options list with initialized fields

This function returns a default options list if called with no arguments, or a pre-formated model option list with checks on the inputs.

stics_wrapper_options(
  javastics_path = NULL,
  stics_exe = "modulostics",
  data_dir = NULL,
  parallel = FALSE,
  cores = NA,
  time_display = FALSE,
  verbose = TRUE,
  force = FALSE,
  successive_usms = NULL,
  ...
)

Arguments

javastics_path
Path of JavaStics installation directory, needed if stics_exe is not provided, or relates to an exe in the javastics_path (see details)

stics_exe
The name, executable or path of the stics executable to use (optional, default to "modulostics", see details)

data_dir
Path(s) of the situation(s) input files directorie(s) or the root path of the situation(s) input files directorie(s)

parallel
Boolean. Is the computation to be done in parallel ?

cores
Number of cores to use for parallel computation.

time_display
Display time

verbose
Logical value (optional), TRUE to display informations during execution, FALSE otherwise (default)

force
Boolean. Don't check javastics_path, stics_exe and data_dir (default to FALSE, see details)

successive_usms
List of vectors containing the names of the UMSs to consider as successive (e.g. list(c("usm1.1","usm1.2"),c("usm2.1","usm2.2")) defines 2 successions usm1.1->usm1.2 and usm2.1->usm2.2)

...
Add further arguments set the options list values

Value

A list containing Stics model stics_wrapper options

Details

stics_exe may be :

a model name pointing to a stics executable as done in JavaStics, e.g. "modulostics" for stics_modulo.exe, the standard version of the model shipping with JavaStics;

a stics executable file available from the bin folder in JavaStics, e.g. "stics_modulo.exe";

a path to a stics executable file, eg. "C:/Users/username/Desktop/stics.exe".

javastics_path must be provided for case (1) and (2).

If force=TRUE, checks are not done for javastics_path, stics_exe and data_dir. In this case, they are returned as is, and will be checked (and potentially updated to match the right stics executable) only at execution of stics_wrapper(). This option is used for portability, when e.g. stics_wrapper_options outputs are sent to a remote.

Examples